Move file logic into main.js
Move the logic on file updates from file_listener.js into main.js in
order to remove dependencies on global variables.
