+++ airtable_object.js
@@ -1,7 +1,7 @@
 "use strict";
 
 let player_cache = new Map();
-exports.get_player_id = async function(name) {
+exports.get_player_id = async function(base, name) {
     if (player_cache.has(name)) {
         return player_cache.get(name);
     }
@@ -21,7 +21,7 @@ exports.get_player_id = async function(name) {
 }
 
 let game_cache = new Map();
-exports.get_game_id = async function(name) {
+exports.get_game_id = async function(base, name) {
     if (game_cache.has(name)) {
         return game_cache.get(name);
     }
@@ -47,7 +47,7 @@ exports.get_game_id = async function(name) {
     @param {string} answer_colum - the colum you want the value of.
     @returns {any}  the value found.
  */
-exports.general_look_up = async function(table, colum_for_search, value, answer_colum) {
+exports.general_look_up = async function(base, table, colum_for_search, value, answer_colum) {
     let records = base(table).select({
         filterByFormula: `{${colum_for_search}} = '${value}'`,
         maxRecords: 1,
