+++ file_listener.js
@@ -1,11 +1,7 @@
 "use strict";
-var Airtable = require('airtable');
 
-const airtable = require("./airtable_object.js")
-const clone = require('rfdc')()
+const events = require('events');
 const fs = require('fs');
-
-global.base = new Airtable({ apiKey: process.env.Api_key }).base(process.env.Base_key);
 
 
 function resolveToAbsolutePath(path) {
@@ -15,6 +11,7 @@ function resolveToAbsolutePath(path) {
 }
 
 exports.watch_files = function(servers) {
+    const file_events = new events.EventEmitter();
     const directories = []
     for (let variable in servers["local_servers"]) {
         const object = servers["local_servers"][variable]
@@ -33,11 +30,15 @@ exports.watch_files = function(servers) {
                 }, 100);
                 var data = readfile(filename, path)
                 console.log(data)
-                run_data(data)
+                var object = JSON.parse(data)
+                file_events.emit(object.type, object);
             }
         });;
     }
+
+    return file_events;
 }
+
 
 function readfile(filename, path) {
     try {
@@ -48,69 +49,4 @@ function readfile(filename, path) {
         return 'Error:' + e.stack
     }
 
-}
-
-
-async function run_data(data) {
-    var object = JSON.parse(data)
-    switch (object.type) {
-        case "Started_game":
-            var json = clone(require("./score_template.json"));
-            var player_ids = []
-            for (let player of object.players) {
-                player_ids.push(await airtable.get_player_id(player))
-            }
-            json[0].fields["Players Present"] = player_ids
-            json[0].fields["Time Started"] = new Date().toISOString();
-            json[0].fields["Game"][0] = await airtable.get_game_id(object.name)
-            console.log("game starting with ")
-            console.log(json)
-            const created = await base('Scoring Data').create(json)
-            console.log(created)
-            global.airtable_id = created[0].id
-            object.arguments.unshift(object.name)
-            console.log(object.arguments)
-            tell_server(object.arguments, object.server, airtable_id)
-            break
-
-        case "end_game":
-            var current_timeDate = new Date()
-            var json = clone(require("./score_update_template.json"));
-            var players
-            json[0].id = airtable_id
-            await Promise.all([airtable.get_player_id(object.Gold), airtable.get_player_id(object.Silver), airtable.get_player_id(object.Bronze)]).then((values) => {
-                players = values
-            });
-            players = players.filter(function(el) {
-                return el != undefined;
-            });
-            json[0].fields["Gold Player"][0] = players[0]
-            json[0].fields["Gold Data"] = object["Gold_data"]
-            if (players[2] != undefined) {
-                json[0].fields["Silver Player"][0] = players[1]
-                json[0].fields["Silver Data"] = object["Silver_data"]
-                if (players[3] != undefined) {
-                    json[0].fields["Bronze Player"][0] = players[2]
-                    json[0].fields["Bronze Data"] = object["Bronze_data"]
-                } else {
-                    delete json[0].fields["Bronze Player"]
-                    delete json[0].fields["Bronze Data"]
-                }
-            } else {
-                delete json[0].fields["Silver Player"]
-                delete json[0].fields["Silver Data"]
-                delete json[0].fields["Bronze Player"]
-                delete json[0].fields["Bronze Data"]
-            }
-            var begin_time = await airtable.gernal_look_up("Scoring Data", "Match ID", airtable_id, "Time Started")
-            begin_time = new Date(begin_time)
-            var difernce = (current_timeDate - begin_time) / 1000 * 60
-            json[0].fields["Duration"] = difernce
-            console.log("game ending with ")
-            console.log(json)
-            send_players(object.server, object)
-            console.log(base('Scoring Data').update(json))
-
-            break
-    }
 }
